pyproject.toml
[project]
name = "sepilot-cli"
version = "0.1.0"
description = "SE Pilot CLI - Intelligent CLI Agent for Software Engineering"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langchain-openai>=0.0.5",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "httpx>=0.25.0",
    "html2text>=2024.0.0",
    "chromadb>=1.3.7",
    "onnxruntime>=1.17,<1.24",
    "chardet<6.0.0",
    "datasets>=2.18.0",
    "swebench>=4.0.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
# Extended LLM providers
llm-providers = [
    "langchain-aws>=0.1.0",
    "langchain-groq>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=0.0.5",
]

# RAG (Retrieval-Augmented Generation) dependencies
rag = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.9.0",
]

# Multimedia (image/PDF) dependencies
multimedia = [
    "Pillow>=10.0.0",
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
]

# Code Intelligence (LSP, Tree-sitter, Indexer)
code-intelligence = [
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-go>=0.21.0",
    "tree-sitter-rust>=0.21.0",
    "watchdog>=3.0.0",
]

# All optional features
all = [
    "sepilot-cli[llm-providers,rag,multimedia,code-intelligence]",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
sepilot = "sepilot.cli.main:main"
sepilot-lsp = "sepilot.cli.main:install_lsp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sepilot"]
